Move NiceFormatter out of utils
This allows us to use it without having to import utils (which pulls in a
lot of other imports)
