+++ src/trex/utils.py
@@ -7,20 +7,6 @@ from .cell import Cell
 from .molecule import Molecule
 
 
-class NiceFormatter(logging.Formatter):
-    """
-    Do not prefix "INFO:" to info-level log messages (but do it for all other
-    levels).
-
-    Based on http://stackoverflow.com/a/9218261/715090 .
-    """
-
-    def format(self, record):
-        if record.levelno != logging.INFO:
-            record.msg = "{}: {}".format(record.levelname, record.msg)
-        return super().format(record)
-
-
 def molecule_list_to_dataframe(molecules: List[Molecule]) -> pd.DataFrame:
     """Convert list of Molecules to pandas DataFrame"""
     molecule_dict = {"cell_id": [], "umi": [], "clone_id": [], "read_count": []}
